게임 로직 구현..(방 설정 반영 x)

+++ src/main/java/kr/ac/ewha/java2/dto/SubmitAnswerRequestDto.java
@@ -9,12 +9,12 @@ public class SubmitAnswerRequestDto {
     //사용자 식별 ID
     private Long participantId;
     //현재 문제의 Id
-    private Integer questionId;
+    private Long questionId;
     //사용자가 제출한 정답 내용
     private String submittedAnswer;
 
     public SubmitAnswerRequestDto(){}
-    public SubmitAnswerRequestDto(Long participantId, Integer questionId, String submittedAnswer){
+    public SubmitAnswerRequestDto(Long participantId, Long questionId, String submittedAnswer){
         this.participantId = participantId;
         this.questionId =  questionId;
         this.submittedAnswer = submittedAnswer;
